pyproject.toml
[project]
name = "they"
version = "0.1.0"
description = "They are just yet another AI Agent."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents[litellm]",
    "rich>=14.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
they = "main:main_sync"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
